+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This changelog covers what's changed in the MPDX APIs.
 
+## 26 April 2017
+- Allowing reverse filters for any contacts or tasks filters.
+    - To do so just add 'reverse_FILTER_NAME_HERE = true' to the filter hash.
+
 ## 20 April 2017
 - Added received_not_processed to pledges.
 - Contact suggested_changes used to sometimes suggest a blank value for status, this was a bug that has been fixed.